Rename LuaScript files to lua_script
