+++ Sprocket/Scripting/lua_input.cpp
@@ -2,7 +2,7 @@
 
 #include <Sprocket/Core/log.h>
 #include <Sprocket/Utility/input_store.h>
-#include <Sprocket/Scripting/LuaScript.h>
+#include <Sprocket/Scripting/lua_script.h>
 #include <Sprocket/Scripting/LuaConverter.h>
 
 #include <lua.h>
